% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{backward_selection}
\alias{backward_selection}
\title{Select covariates backwise}
\usage{
backward_selection(
  model,
  threshold = 0.15,
  measure_fn = function(x) summary(x)[["coefficients"]][, 4],
  max_steps = 1000,
  return_step_results = FALSE,
  do_not_remove = c("(Intercept)"),
  ...
)
}
\arguments{
\item{model}{A model with [stats::update()], [stats::coef()] methods.}

\item{threshold}{Value threshold to remove variable. It can be a fixed value
or a function. The variable is removed if `measure_fn(model) > threshold` and
added if `measure_fn(model) <= threshold`.}

\item{measure_fn}{Function with model as argument and returns values to be used by
`threshold`. It can also compare two models, where during forward step
it calls `measure_fn(candidate_model, current_selected_model)` and
during backward step it calls `measure_fn(current_selected_model, candidate_model)`.
Defaults to the p-value from the summary of the coefficients.}

\item{max_steps}{The maximum number of steps for the variable selection process.
Defaults to 1000.}

\item{return_step_results}{Logical. If TRUE, the function returns a list
containing the final fitted model and a log of the selection steps.
Defaults to FALSE.}

\item{do_not_remove}{A character vector specifying variables that should not
be removed during backward selection. Defaults to "(Intercept)".}

\item{...}{Extra arguments to [stats::update()].}
}
\value{
the selected model if `return_step_results` is `False`.
}
\description{
Select covariates backwise
}
\examples{
\dontrun{
model <- lm(mpg ~ ., data = mtcars)
backward_selection(model)
}

}
