% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{compute_statistic}
\alias{compute_statistic}
\title{Compute Wald Statistic}
\usage{
compute_statistic(coefs, vcov, generator_coef)
}
\arguments{
\item{coefs}{coefficient vector.}

\item{vcov}{Covariance Matrix, should be a square matrix with p rows.}

\item{generator_coef}{coefficient vector used to generate `coefs` and `vcov`.}
}
\value{
Wald statistic for each coefficient.
}
\description{
Compute Wald statistic for one coefficient using the simulation coefficient
and standard error.
}
