% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat.R
\name{concat_pvalues}
\alias{concat_pvalues}
\title{Concatenate LD_pvalues object}
\usage{
concat_pvalues(ld_pvalues)
}
\arguments{
\item{ld_pvalues}{list of elements of class \code{LD_pvalues}.}
}
\value{
Object of class \code{LD_pvalues}
}
\description{
Concatenate LD_pvalues object
}
