% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.R
\name{envelope_measures}
\alias{envelope_measures}
\title{Generate Simulated Envelope Measures}
\usage{
envelope_measures(
  model,
  residual_fn = stats::rstudent,
  alpha = 0.05,
  n_sim = 100,
  ...
)
}
\arguments{
\item{model}{A model that has a \code{\link[stats:simulate]{stats::simulate()}} method and is compatible with \code{\link[=get_refit]{get_refit()}}.}

\item{residual_fn}{A function to calculate model residuals. The default is
\code{\link[stats:influence.measures]{stats::rstudent()}} for studentized residuals.}

\item{alpha}{The significance level for constructing the envelope bounds.
Defaults to 0.05.}

\item{n_sim}{The number of simulations to perform for envelope construction.
Defaults to 100.}

\item{...}{Extra arguments to \code{\link[=get_refit]{get_refit()}}}
}
\description{
Generate Simulated Envelope Measures
}
\examples{
\dontrun{
fit <- lm(mpg ~ cyl, data = mtcars)

envelope_measures(fit, residual_fn = function(x) residuals.lm(x, type = "deviance"))

envelope_measures(fit, n_sim = 200)
envelope_measures(fit, residual_fn = rstandard, n_sim = 200)
}

}
\seealso{
\code{\link{get_refit}}, \code{\link{simulate}}, \code{\link{rstudent}}
}
