% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{get_p_values_matrix}
\alias{get_p_values_matrix}
\title{Generate P-Values Matrix from Model Coefficient Simulations}
\usage{
get_p_values_matrix(model, generator, n_sim = 1000, df = NULL, ...)
}
\arguments{
\item{model}{A model with \code{\link[=update]{update()}}, \code{\link[=simulate]{simulate()}}, \code{\link[=fitted]{fitted()}},
\code{\link[=coef]{coef()}}, \code{\link[=vcov]{vcov()}} methods.}

\item{generator}{An optional function 2 arguments (n, mu) to generate new response vectors.
If NULL, the model's \code{\link[=simulate]{simulate()}} method is used. Otherwise, the generator
function is used to simulate responses.}

\item{n_sim}{The number of simulations to perform.}

\item{df}{The degrees of freedom for the t-test statistic. If NULL, use
Chisq-test statistic.}

\item{...}{Additional arguments to be passed to \code{simulate_coefficients}.}
}
\value{
A matrix where each column represents the p-values obtained from a simulation.
}
\description{
This function generates a matrix of p-values by simulating coefficients from
a given model and computing test statistics for each simulation.
}
\examples{
\dontrun{
model <- lm(mpg ~ wt + hp, data = mtcars)
generator <- function(n, mu) {
  rt(n, 3) + mu
}
get_p_values_matrix(model, generator, n_sim = 100)
}

}
