% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.R
\name{plot.LD_envelope}
\alias{plot.LD_envelope}
\title{Envelope Plot}
\usage{
\method{plot}{LD_envelope}(
  x,
  colors = c("red", "black"),
  xlab = "Expected quantiles",
  ylab = "Observed quantiles",
  ...
)
}
\arguments{
\item{x}{LD_envelope object, usually the result of \code{\link[=envelope]{envelope()}}}

\item{colors}{Vector of length 2, with color for points outside and inside
the envelope band, respectivelly.}

\item{xlab}{The label for the x-axis.}

\item{ylab}{The label for the y-axis.}

\item{...}{extra arguments passed to \link[graphics:plot.default]{graphics::plot}}
}
\description{
Plot LD_envelope
}
