% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{plot.LD_pvalues}
\alias{plot.LD_pvalues}
\title{Plot Joint Empirical Cumulative Distribution Function (ECDF) of p-values}
\usage{
\method{plot}{LD_pvalues}(
  x,
  which = seq_len(length(x$test_coefficients) + 1),
  caption = as.list(paste("ECDF of", c(names(x$test_coefficients), "all coefficients"))),
  plot_uniform = TRUE,
  uniform_legend = TRUE,
  ylab = "Empirical cumulative distribution",
  xlab = "p-value",
  ...,
  ask = prod(graphics::par("mfcol")) < length(which) && grDevices::dev.interactive()
)
}
\arguments{
\item{x}{LD_pvalues object, usually the result of \code{\link[=get_p_values]{get_p_values()}}}

\item{which}{A vector specifying the indices of coefficients to plot.
If index is bigger than the number of coefficients it plots the joint p_value.}

\item{caption}{A list with caption for each plot.}

\item{plot_uniform}{Logical. If TRUE, plot uniform distribution.}

\item{uniform_legend}{Logical. If TRUE, a legend is added to the plot to
distinguish between the p-value and U(0, 1) curves. Defaults to TRUE.}

\item{ylab}{The label for the y-axis. Defaults to "Empirical cumulative distribution".}

\item{xlab}{The label for the x-axis. Defaults to "p-value".}

\item{...}{extra arguments passed to \link[graphics:plot.default]{graphics::plot}}

\item{ask}{Logical. If TRUE, the user is prompted before each plot. Defaults
to TRUE if in an interactive session and the number of plots is greater
than the available space; otherwise, FALSE.}
}
\value{
A vector of joint p-values for all coefficients.
}
\description{
Plot Joint Empirical Cumulative Distribution Function (ECDF) of p-values
}
\examples{
model <- lm(mpg ~ wt + hp, data = mtcars)
p_values_ld <- get_p_values(model, n_sim = 100)
plot(p_values_ld)
}
