% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{plot_cook}
\alias{plot_cook}
\title{Plot Cook's distances}
\usage{
plot_cook(
  model,
  n_highlights = 0,
  cut = FALSE,
  xlab = "Index",
  ylab = "Cook's distance",
  ...
)
}
\arguments{
\item{model}{Model with [cooks.distance()] method}

\item{n_highlights}{Number of points with highest cook distance to highlight in the plot}

\item{cut}{Boolean indicating to show horizontal line of 4 standard deviations
above the Cook's distances mean.}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{...}{Further arguments for [graphics::plot()]}
}
\value{
Cook's distances invisible
}
\description{
Plot Cook's distances
}
\examples{
fit <- lm(mpg ~ cyl, data = mtcars)

plot_cook(fit)
plot_cook(fit, n_highlights = 2)
}
