% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{plot_pvalues_ecdf}
\alias{plot_pvalues_ecdf}
\title{Plot Simulation p-values}
\usage{
plot_pvalues_ecdf(
  model,
  generator = NULL,
  n_sim = 1000,
  which = seq_along(stats::coef(model)),
  caption = paste("ECDF of", names(stats::coef(model))[which]),
  plot_uniform = TRUE,
  uniform_legend = TRUE,
  ylab = "Empirical cumulative distribution",
  xlab = "p-value",
  args_sim = list(),
  ...,
  ask = prod(graphics::par("mfcol")) < length(which) && grDevices::dev.interactive(),
  use_tstat = NULL
)
}
\arguments{
\item{model}{A model with [update()] method.}

\item{generator}{Optional custom generator function with 2 arguments (n, mu).}

\item{n_sim}{Number of simulations.}

\item{which}{vector of integers with model coefficients indices to plot.}

\item{caption}{Plot title for each coefficient.}

\item{plot_uniform}{Boolean to plot uniform distribution.}

\item{uniform_legend}{Boolean show legend indicating uniform line
when `plot_uniform` is `TRUE`.}

\item{ylab}{y-axis label}

\item{xlab}{x-axis label}

\item{args_sim}{Extra arguments passed to [simulate_coefficients()].}

\item{...}{Extra arguments from [plot()]}

\item{ask}{Logical, ask to show next plot}

\item{use_tstat}{Logical, Should use t-test using model df.residual.
If `NULL` look at the model summary to decide.}
}
\value{
Matrix with p_values obtained from the simulations.
}
\description{
Plot Simulation p-values
}
\examples{
fit <- lm(mpg ~ cyl, data = mtcars)

plot_pvalues_ecdf(fit)
}
