% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{plot_res_vs_linear_predictor}
\alias{plot_res_vs_linear_predictor}
\title{Plot Residuals against Linear Predictor}
\usage{
plot_res_vs_linear_predictor(
  model,
  residual_fn = stats::rstandard,
  xlab = "Linear Predictor",
  ylab = "Standardized deviance residuals",
  ...
)
}
\arguments{
\item{model}{Model with methods [predict()] or [fitted()]}

\item{residual_fn}{function to compute residuals, default [rstandard()],
other functions could be [rstudent()], [residuals()].}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{...}{Extra arguments to `residual_fn`}
}
\value{
invisible list with plot components
}
\description{
Plot Residuals against Linear Predictor
}
\details{
If the model was fitted using the [glm()] function, it will use the [predict()]
method, otherwise, it will use the [fitted()] method.
}
\examples{
fit <- lm(mpg ~ cyl, data = mtcars)

plot_res_vs_linear_predictor(fit)
plot_res_vs_linear_predictor(fit, residual_fn = rstudent)
plot_res_vs_linear_predictor(fit, residual_fn = residuals)


glm_fit <- glm(cyl ~ mpg, family = poisson(), data = mtcars)

plot_res_vs_linear_predictor(glm_fit)
plot_res_vs_linear_predictor(glm_fit, type = "pearson")
}
