% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_refit.R
\name{update_using_formula}
\alias{update_using_formula}
\title{Update the object using only the formula}
\usage{
update_using_formula(object, new_response, ...)
}
\arguments{
\item{object}{A model.}

\item{new_response}{the new response, may be a vector or a matrix.}

\item{...}{other arguments passed to \code{refit} or \code{update}.}
}
\description{
Create a new formula with \code{new_response} values directly on the formula left hand side.
}
\details{
This function is designed to be used as a fallback for \code{\link[=update_using_model_frame]{update_using_model_frame()}}
as it's prone to run with memory issues.
}
