% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_refit.R
\name{update_using_model_frame}
\alias{update_using_model_frame}
\title{Update the object with new response using only model frame.}
\usage{
update_using_model_frame(object, new_response, ...)
}
\arguments{
\item{object}{A model.}

\item{new_response}{the new response, may be a vector or a matrix.}

\item{...}{other arguments passed to \code{refit} or \code{update}.}
}
\description{
Update the object with new response using only model frame.
}
