% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{backward_selection}
\alias{backward_selection}
\title{Select covariates backwise}
\usage{
backward_selection(
  model,
  threshold = 0.15,
  measure_fn = function(x) summary(x)[["coefficients"]][, 4],
  data = NULL,
  max_steps = 1000,
  return_step_results = FALSE,
  do_not_remove = c("(Intercept)")
)
}
\arguments{
\item{model}{A model with [stats::update()] method.}

\item{threshold}{Value threshold to remove variable. where the variable is
removed if `measure_fn(model) >= threshold`.}

\item{measure_fn}{Function with model as argument and returns values to be used by
`threshold`. Default is p-value reported by [summary()].}

\item{data}{Data to be used for model refit.}

\item{max_steps}{Maximum number of steps for selection.}

\item{return_step_results}{Boolean that if `TRUE` returns a list with the first
argument the selection result, and second argument a list with step, operation,
variable and its value.
`measure_fn` measures and coefficient removed.}

\item{do_not_remove}{String vector of variables to not be removed in the selection.}
}
\value{
the selected model if `return_step_results` is `False`.
}
\description{
Select covariates backwise
}
\examples{
model <- lm(mpg ~ ., data = mtcars)
backward_selection(model)
}
