% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.R
\name{envelope_measures}
\alias{envelope_measures}
\title{Generate Simulated Envelope Measures}
\usage{
envelope_measures(
  model,
  residual_fn = stats::rstudent,
  alpha = 0.05,
  n_sim = 100,
  ...
)
}
\arguments{
\item{model}{A model with \code{\link[stats:update]{stats::update()}} and \code{\link[stats:simulate]{stats::simulate()}} methods.}

\item{residual_fn}{A function to calculate model residuals. The default is
\code{stats::rstudent} for studentized residuals.}

\item{alpha}{The significance level for constructing the envelope bounds.
Defaults to 0.05.}

\item{n_sim}{The number of simulations to perform for envelope construction.
Defaults to 100.}

\item{...}{Extra arguments to \code{\link[stats:update]{stats::update()}}}
}
\value{
A list containing the following components:
\describe{
\item{expected}{A vector of expected quantiles from a normal distribution.}
\item{observed}{A vector of observed quantiles from the model residuals.}
\item{outside}{A logical vector indicating whether each observation falls
outside the constructed envelope bounds.}
\item{lower}{The lower bounds of the envelope for each observation.}
\item{med}{The median bounds of the envelope for each observation.}
\item{upper}{The upper bounds of the envelope for each observation.}
}
}
\description{
Generate Simulated Envelope Measures
}
\examples{
\dontrun{
fit <- lm(mpg ~ cyl, data = mtcars)

envelope_measures(fit, residual_fn = function(x) residuals.lm(x, type = "deviance"))

envelope_measures(fit, n_sim = 200)
envelope_measures(fit, residual_fn = rstandard, n_sim = 200)
}

}
\seealso{
\code{\link{update}}, \code{\link{simulate}}, \code{\link{rstudent}}
}
