% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.R
\name{envelope_measures}
\alias{envelope_measures}
\title{Generate Simulated Envelope Measures}
\usage{
envelope_measures(
  model,
  residual_fn = stats::rstudent,
  alpha = 0.05,
  n_sim = 100
)
}
\arguments{
\item{model}{A model with `update` method}

\item{residual_fn}{Function to compute residual using `model`}

\item{alpha}{Confidence Band level}

\item{n_sim}{Number of simulations}
}
\value{
A list of vectors with the expected, observed quantiles,
boolean, indicating inside or outside of band; and
lower band values with `alpha/2`, median, upper band values with `1 - alpha/2`.
}
\description{
Generate Simulated Envelope Measures
}
\examples{
fit <- lm(mpg ~ cyl, data = mtcars)

envelope_measures(fit, residual_fn = function(x) residuals.lm(x, type = "deviance"))

envelope_measures(fit, n_sim = 200)
envelope_measures(fit, residual_fn = rstandard, n_sim = 200)
}
