% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{get_p_values_joint}
\alias{get_p_values_joint}
\title{Generate P-Values from Model Coefficient Simulations}
\usage{
get_p_values_joint(model, generator, n_sim = 1000, ...)
}
\arguments{
\item{model}{A model with [update()], [simulate()], [fitted()],
[coef()], [vcov()] methods.}

\item{generator}{An optional function 2 arguments (n, mu) to generate new response vectors.
If NULL, the model's [simulate()] method is used. Otherwise, the generator
function is used to simulate responses.}

\item{n_sim}{The number of simulations to perform.}

\item{...}{Additional arguments to be passed to `simulate_coefficients`.}
}
\value{
A numeric vector containing the joint p-values obtained from each simulation.
}
\description{
This function generates p-values by simulating coefficients from
a given model and computing Wald statistics for each simulation.
}
\examples{
\dontrun{
model <- lm(mpg ~ wt + hp, data = mtcars)
get_p_values_joint(model, n_sim = 100)
}

}
