% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.R
\name{plot_envelope}
\alias{plot_envelope}
\title{Plot Simulated Envelope}
\usage{
plot_envelope(
  model,
  xlab = "Expected quantiles",
  ylab = "Observed quantiles",
  ...
)
}
\arguments{
\item{model}{A model with `update` method}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{...}{Additional arguments for `envelope_measures()`}
}
\value{
Creates the simulated envelope plot and returns its components
}
\description{
Plot Simulated Envelope
}
\examples{
fit <- lm(mpg ~ cyl, data = mtcars)

plot_envelope(fit)
}
