% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.R
\name{plot_envelope}
\alias{plot_envelope}
\title{Plot Simulated Envelope}
\usage{
plot_envelope(
  model,
  colors = c("red", "black"),
  xlab = "Expected quantiles",
  ylab = "Observed quantiles",
  ...
)
}
\arguments{
\item{model}{A model that has a \code{\link[stats:simulate]{stats::simulate()}} method and is compatible with \code{\link[=get_refit]{get_refit()}}.}

\item{colors}{A vector with two strings, the first element is the color of the
elements outside of the envelope, the second is for the elements inside.}

\item{xlab}{The label for the x-axis. Defaults to "Expected quantiles".}

\item{ylab}{The label for the y-axis. Defaults to "Observed quantiles".}

\item{...}{Additional arguments for \code{\link[=envelope_measures]{envelope_measures()}}}
}
\value{
A list containing the following components:
\describe{
\item{expected}{A vector of expected quantiles from a normal distribution.}
\item{observed}{A vector of observed quantiles from the model residuals.}
\item{outside}{A logical vector indicating whether each observation falls
outside the constructed envelope bounds.}
\item{lower}{The lower bounds of the envelope for each observation.}
\item{med}{The median bounds of the envelope for each observation.}
\item{upper}{The upper bounds of the envelope for each observation.}
}
}
\description{
This function generates a plot to visualize the envelope measures.
The plot includes expected quantiles versus observed
quantiles, along with lower, median, and upper bounds of the constructed
envelope.
}
\details{
This function calls \code{\link[=envelope_measures]{envelope_measures()}} and plot it's results.
}
\examples{
\dontrun{
model <- lm(mpg ~ wt + hp, data = mtcars)
plot_envelope(model)
}

}
\seealso{
\code{\link[graphics:plot.default]{graphics::plot()}}
}
