% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{plot_joint_pvalues_ecdf}
\alias{plot_joint_pvalues_ecdf}
\title{Plot Joint Empirical Cumulative Distribution Function (ECDF) of p-values}
\usage{
plot_joint_pvalues_ecdf(
  model,
  plot_uniform = TRUE,
  uniform_legend = TRUE,
  ylab = "Empirical cumulative distribution",
  xlab = "p-value",
  args_sim = list(),
  ...
)
}
\arguments{
\item{model}{A model with \code{\link[=update]{update()}}, \code{\link[=simulate]{simulate()}}, \code{\link[=fitted]{fitted()}},
\code{\link[=coef]{coef()}}, \code{\link[=vcov]{vcov()}} methods.}

\item{plot_uniform}{Logical. If TRUE, plot uniform distribution.}

\item{uniform_legend}{Logical. If TRUE, a legend is added to the plot to
distinguish between the p-value and U(0, 1) curves. Defaults to TRUE.}

\item{ylab}{The label for the y-axis. Defaults to "Empirical cumulative distribution".}

\item{xlab}{The label for the x-axis. Defaults to "p-value".}

\item{args_sim}{Extra arguments passed to \code{\link[=get_p_values_joint]{get_p_values_joint()}}.}

\item{...}{Extra arguments from \code{\link[=plot]{plot()}}}
}
\value{
A vector of joint p-values for all coefficients.
}
\description{
This function generates a plot displaying the joint empirical cumulative
distribution function (ECDF) of p-values for all coefficients in a model.
The p-values are computed based on simulated coefficients and covariance
matrices.
}
\examples{
\dontrun{
fit <- lm(mpg ~ cyl, data = mtcars)
plot_joint_pvalues_ecdf(fit)
}
}
