% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{plot_pvalues_ecdf}
\alias{plot_pvalues_ecdf}
\title{Plot Empirical Cumulative Distribution Function (ECDF) of p-values}
\usage{
plot_pvalues_ecdf(
  model,
  which = seq_along(stats::coef(model)),
  caption = paste("ECDF of", names(stats::coef(model))[which]),
  plot_uniform = TRUE,
  uniform_legend = TRUE,
  ylab = "Empirical cumulative distribution",
  xlab = "p-value",
  args_sim = list(),
  ...,
  ask = prod(graphics::par("mfcol")) < length(which) && grDevices::dev.interactive(),
  use_tstat = NULL
)
}
\arguments{
\item{model}{A model with \code{\link[=update]{update()}}, \code{\link[=simulate]{simulate()}}, \code{\link[=fitted]{fitted()}},
\code{\link[=coef]{coef()}}, \code{\link[=vcov]{vcov()}} methods.}

\item{which}{A vector specifying the indices of coefficients to plot. Defaults
to all coefficients.}

\item{caption}{A character vector providing plot captions for each coefficient.}

\item{plot_uniform}{Logical. If TRUE, plot uniform distribution.}

\item{uniform_legend}{Logical. If TRUE, a legend is added to the plot to
distinguish between the p-value and U(0, 1) curves. Defaults to TRUE.}

\item{ylab}{The label for the y-axis. Defaults to "Empirical cumulative distribution".}

\item{xlab}{The label for the x-axis. Defaults to "p-value".}

\item{args_sim}{Extra arguments passed to \code{\link[=simulate_coefficients]{simulate_coefficients()}}.}

\item{...}{Extra arguments from \code{\link[=plot]{plot()}}}

\item{ask}{Logical. If TRUE, the user is prompted before each plot. Defaults
to TRUE if in an interactive session and the number of plots is greater
than the available space; otherwise, FALSE.}

\item{use_tstat}{Logical. If TRUE, the t-statistic is used for computing p-values.
If FALSE, the z-statistic is used. If NULL, the default is determined based on
the presence of the "t value" column in the summary of the model coefficients.}
}
\value{
Matrix with p_values obtained from the simulations.
}
\description{
This function generates a series of plots displaying the empirical cumulative
distribution function (ECDF) of p-values for selected coefficients in a
model. The p-values are computed based on simulated coefficients
and covariance matrices.
}
\examples{
\dontrun{
fit <- lm(mpg ~ cyl, data = mtcars)

plot_pvalues_ecdf(fit)
}
}
