% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{simulate_coefficients}
\alias{simulate_coefficients}
\title{Simulate Model Coefficients and Covariance Matrix}
\usage{
simulate_coefficients(model, generator = NULL, ...)
}
\arguments{
\item{model}{A model with \code{\link[=update]{update()}}, \code{\link[=simulate]{simulate()}}, \code{\link[=fitted]{fitted()}},
\code{\link[=coef]{coef()}}, \code{\link[=vcov]{vcov()}} methods.}

\item{generator}{An optional function 2 arguments (n, mu) to generate new response vectors.
If NULL, the model's \code{\link[=simulate]{simulate()}} method is used. Otherwise, the generator
function is used to simulate responses.}

\item{...}{Extra arguments to \code{\link[stats:update]{stats::update()}}}
}
\value{
A list containing the simulated coefficients and covariance matrices.
The list includes components:
\describe{
\item{coefs}{Coefficient vectors.}
\item{vcov}{Covariance matrices.}
}
}
\description{
This function simulates coefficients and the covariance matrix for a fitted
model. It allows for generating new response vectors either using
the model's simulation method or a user-specified generator function.
}
\examples{
\dontrun{
fit <- lm(mpg ~ cyl, data = mtcars)
generator <- function(n, mu) rt(n, 3) + mu
simulate_coefficients(fit, generator = generator)
}

}
