% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{simulate_coefficients}
\alias{simulate_coefficients}
\title{Simulate Model Coefficients}
\usage{
simulate_coefficients(model, generator = NULL, n_sim = 100)
}
\arguments{
\item{model}{A model with `update` method}

\item{generator}{Optional custom generator function with 2 arguments (n, mu)}

\item{n_sim}{Number of simulations}
}
\value{
A list with a list for coefficients and a list with vcov, each of length n_sim
}
\description{
Simulate Model Coefficients
}
\examples{
fit <- lm(mpg ~ cyl, data = mtcars)

generator <- function(n, mu) rt(n, 3) + mu

simulate_coefficients(fit, generator = generator)

library(MASS)

fit <- glm.nb(Days ~ Sex/(Age + Eth*Lrn), data = quine)

simulate_coefficients(fit)
}
